refactor: DSL 실습 코드 분리
