+++ src/main/kotlin/dslstudy/Person.kt
@@ -1,0 +1,8 @@
+package dslstudy
+
+data class Person(
+    val name: String,
+    val company: String?,
+    val skill: Skill?,
+    val languages: Map<String, Int>?,
+)
