+++ src/main/kotlin/dslstudy/Skill.kt
@@ -1,0 +1,6 @@
+package dslstudy
+
+data class Skill(
+    val softSkill: List<String>,
+    val hardSkill: List<String>,
+)
